# spring-server/src/main/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: Articles app API 
  description: "API de mon grand projet qui n'a pas encore de nom. App sur laquelle on peut poster des articles et intéragir avec ceux-ci."
  version: 1.0.0
servers:
- url: https://localhost:8080/api
...
paths:
...
  /users:
    post:
      tags:
      - users
      summary: Create user
      description: creates a user
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        default:
          description: successful operation
          content: {}
  /users/search:
    get:
      tags:
      - users
      summary: search for a user
      operationId: searchUser
      parameters:
      - name: search
        in: query
        description: The word for the research
        required: true
        schema: 
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/register:
    post:
      tags: 
        - users
      summary: Registers a user into the system
      operationId: registerUser
      requestBody:
        description: infos about the new user
        content:
          '*/*':
            schema: 
              $ref: '#/components/schemas/User'
      responses: 
        200: 
          description: a new User was correctly added
          content: {}
  /users/login:
    post:
      tags:
      - users
      summary: Logs a user into the system
      operationId: loginUser
      requestBody:
        description: infos about the user to log in
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: successful operation
        400:
          description: Invalid username/password supplied
          content: {}
...
  /users/{userId}:
    get:
      tags:
      - users
      summary: Get user by user name
      operationId: getUserByName
      parameters:
      - name: userId
        in: path
        description: 'The name that needs to be fetched.'
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid username supplied
          content: {}
        404:
          description: User not found
          content: {}
    patch:
      tags:
      - users
      summary: Update a user by ID
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        description: name that need to be updated
        required: true
        schema:
          type: string
      requestBody:
        description: Updated user object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        400:
          description: Invalid user supplied
          content: {}
        404:
          description: User not found
          content: {}
    delete:
      tags:
      - users
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        description: The name that needs to be deleted
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid username supplied
          content: {}
        404:
          description: User not found
          content: {}
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
...